+++ apps/user/src/shared/components/notApproved/style/style.css.ts
