+++ apps/user/src/shared/types/declarations.d.ts
