+++ apps/user/src/shared/types/svgr.d.ts
