+++ apps/user/src/shared/types/user.d.ts
